backend/requirements.txt
# FastAPI et serveur
fastapi==0.115.8
uvicorn[standard]==0.34.0

# Configuration
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Base de données
sqlmodel==0.0.22
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis==5.2.1

# Celery
celery==5.4.0

# Sécurité
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Rate Limiting
slowapi==0.1.9

# Compression
brotli-asgi==1.4.0

# Utilitaires
httpx==0.28.1
structlog==25.1.0

# Extraction de texte
pdfplumber==0.11.5
python-docx==1.1.2

# Tests
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-cov==6.0.0
httpx==0.28.1

# Qualité du code (dev)
flake8==7.1.1
mypy==1.14.1
black==24.10.0
isort==5.13.2
